# docs/openapi/transfers.yaml
openapi: 3.0.3
info:
  title: Bank Transfer API
  version: 1.0.0

paths:
  /v1/transfers:
    post:
      summary: Transfer funds between two accounts
      operationId: createTransfer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        "200":
          description: Transfer completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    TransferRequest:
      type: object
      required:
        - from_account_id
        - to_account_id
        - amount
      properties:
        from_account_id:
          type: string
        to_account_id:
          type: string
        amount:
          type: integer
          minimum: 1
    ErrorResponse:
      type: object
      properties:
        message:
          type: string